Issue #17 - Move config file from utils to own file (and fix minor bugs)
